Redo class period student management
It used to just have a single update method. This was problematic
because you really want to warn a teacher before removing a student from
a class because they'll lose all of that student's records for that
class when then do. Allowing them to remove students in bulk made it
awkward to give that warning.

The flow is still not amazing, but it's better.